Servico de animals implementado

+++ FarmManagementSystem.Services/Services/AnimalService.cs
@@ -1,11 +1,13 @@
 ﻿using FarmManagementSystem.Domain.Entities;
 using FarmManagementSystem.Domain.Interfaces.IRepositories;
+using System.ComponentModel.DataAnnotations;
 
 namespace FarmManagementSystem.Services.Services
 {
-    public class AnimalService(IAnimalRepository animalRepository)
+    public class AnimalService(IAnimalRepository animalRepository, IFarmRepository farmRepository)
     {
         private readonly IAnimalRepository _animalRepository = animalRepository;
+        private readonly IFarmRepository _farmRepository = farmRepository;
 
         public List<Animal> GetAll()
         {
@@ -23,18 +25,28 @@ namespace FarmManagementSystem.Services.Services
         {
             try
             {
-                return _animalRepository.GetById(Id);
+                var animal = _animalRepository.GetById(Id);
+
+                if (animal == null)
+                    throw new ValidationException("Não existem registros desse animal em nosso sistemma.");
+
+                return animal;
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                throw new Exception("Animal não encontrado");
+                throw new Exception(ex.Message);
             }
         }
 
-        public List<Animal> GetByFarmId(int farmId)
+        public Animal GetByFarmId(int farmId)
         {
             try
             {
+                var animal = _animalRepository.GetByFarmId(farmId);
+
+                if (animal == null)
+                    throw new ValidationException("Não existem registros desse animal em nosso sistemma.");
+
                 return _animalRepository.GetByFarmId(farmId);
             }
             catch (Exception ex)
@@ -47,6 +59,7 @@ namespace FarmManagementSystem.Services.Services
         {
             try
             {
+                animal.Validate();
                 _animalRepository.Add(animal);
             }
             catch (Exception ex)
@@ -59,7 +72,19 @@ namespace FarmManagementSystem.Services.Services
         {
             try
             {
+                animal.Validate();
+                animal.ValidateId();
+
                 var animalInDb = _animalRepository.GetById(animal.Id);
+
+                if (animalInDb == null)
+                    throw new ValidationException("Não existem registros desse animal em nosso sistemma.");
+
+                var farm = _farmRepository.GetById(animalInDb.FarmId);
+
+                if (!farm.IsFarmActive())
+                    throw new ValidationException("A fazenda que contem esse animal está inativa.");
+
                 _animalRepository.Update(animalInDb, animal);
             }
             catch (Exception ex)
@@ -73,6 +98,15 @@ namespace FarmManagementSystem.Services.Services
             try
             {
                 var animalInDb = _animalRepository.GetById(id);
+
+                if (animalInDb == null)
+                    throw new ValidationException("Não existem registros desse animal em nosso sistemma.");
+
+                var farm = _farmRepository.GetById(animalInDb.FarmId);
+
+                if (!farm.IsFarmActive())
+                    throw new ValidationException("A fazenda que contem esse animal está inativa.");
+
                 _animalRepository.Delete(animalInDb);
             }
             catch (Exception ex)
